# docs/swagger.yaml
basePath: /api/v1/
definitions:
  httpresponse.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  urlroute.CreateURLAliasRequest:
    properties:
      original_url:
        type: string
    type: object
  urlroute.CreateURLAliasResponse:
    properties:
      alias:
        type: string
    type: object
  urlroute.GetOriginalByAliasResponse:
    properties:
      original_url:
        type: string
    type: object
info:
  contact:
    name: API [Roman] Support
  description: Swagger API for Golang Project URL Shortener.
  license:
    name: romandnk
    url: https://github.com/romandnk/shortener
  termsOfService: http://swagger.io/terms/
  title: URL shortener project
  version: "1.0"
paths:
  /urls:
    post:
      description: Create short new URL alias if not exists.
      parameters:
      - description: Required JSON body with original url
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/urlroute.CreateURLAliasRequest'
      responses:
        "201":
          description: URL alias was created successfully
          schema:
            $ref: '#/definitions/urlroute.CreateURLAliasResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/httpresponse.Response'
      summary: Create short URL alias
      tags:
      - URL
  /urls/:alias:
    get:
      description: Get original URL by its alias.
      parameters:
      - description: Required path param with original url alias
        in: path
        name: alias
        required: true
        type: string
      responses:
        "200":
          description: Short URL was received successfully
          schema:
            $ref: '#/definitions/urlroute.GetOriginalByAliasResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/httpresponse.Response'
      summary: Get original URL
      tags:
      - URL
swagger: "2.0"